Latest Changes

+++ lib/QueryProvider/query_provider.dart
@@ -2082,7 +2082,7 @@ class QueryProvider {
       token,rating, feedback, bookingId, bookingType) async {
     String _query = """ 
     mutation {
-        addRating(rating: ${int.parse(rating.toString())}, feedback: "$feedback", bookingId: ${int.parse(bookingId.toString())},  bookingType: 1) {
+        addRating(rating: ${double.parse(rating.toString())}, feedback: "$feedback", bookingId: ${int.parse(bookingId.toString())},  bookingType: 1) {
           message
         }
       }
